feat(registry): consolidate all family landings under /families
Move the AI hub to /families/ai (a Next static segment that wins over the
dynamic [category]) so every family lives under one namespace, and add a
/families index listing all 12 families. familyPath now returns
/families/<category> for all; old /ai permanently redirects to /families/ai
(preserves SEO). Unify the template: family pages get the Footer like the AI
page. Update internal links (build, vs, component breadcrumb, landing) and the
sitemap (drop standalone /ai, add /families + /families/ai).
